+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 * Update to wikimedia/zest-css 2.0.0.
 * Update to wikimedia/idle-dom 0.8.0.
 * Implement the following non-standard methods for PHP compatibility:
+  * Document::loadHTML()
   * Document::loadXML()
+  * Document::saveHTML()
   * Document::saveXML()
   * DocumentFragment::appendXML()
   * Element::setIdAttribute() (stubbed out)